refactor: move component to shell
