+++ shell/components/__tests__/Cron/CronExpressionEditorModal.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable jest/no-hooks */
 import { mount, VueWrapper } from '@vue/test-utils';
 import { createStore } from 'vuex';
-import CronExpressionEditorModal from '@components/Cron/CronExpressionEditorModal.vue';
+import CronExpressionEditorModal from '@shell/components/Cron/CronExpressionEditorModal.vue';
 
 interface CronExpressionEditorModalVm extends InstanceType<typeof CronExpressionEditorModal> {
   localCron: string;
